Use PyTorchs.to() function for Tensor dtype conversion.

+++ segment_anything/predictor.py
@@ -161,8 +161,8 @@ class SamPredictor:
         )
 
         masks_np = masks[0].detach().cpu().numpy()
-        iou_predictions_np = iou_predictions[0].detach().cpu().numpy()
-        low_res_masks_np = low_res_masks[0].detach().cpu().numpy()
+        iou_predictions_np = iou_predictions[0].detach().cpu().to(torch.float32).numpy()
+        low_res_masks_np = low_res_masks[0].detach().cpu().to(torch.float32).numpy()
         return masks_np, iou_predictions_np, low_res_masks_np
 
     @torch.no_grad()
